pyproject.toml
[tool.poetry]
name = "pneumonia-detection"
version = "0.1.0"
description = "Chest X-ray pneumonia detection using PyTorch Lightning"
authors = ["Aleksey Sanin <sanin.aa@phystech.edu>"]
readme = "README.md"
packages = [{ include = "pneumonia_detection" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

torch = "2.2.2"
torchvision = "0.17.2"

pytorch-lightning = ">=2.2.0,<2.3.0"
torchmetrics = ">=1.3.0,<2.0.0"

hydra-core = ">=1.3.2,<1.4.0"
omegaconf = ">=2.3.0,<2.4.0"

mlflow = ">=2.13.0,<3.0"
dvc = ">=3.34.0,<4.0.0"

onnx = ">=1.15.0,<2.0.0"
onnxruntime = "1.18.0"

pillow = ">=10.0.0,<11.0.0"
numpy = ">=1.24.0,<2.0.0"
scikit-learn = ">=1.3.0,<2.0.0"
matplotlib = ">=3.7.0,<4.0.0"
fire = "^0.5.0"
setuptools = ">=65,<70"

kaggle = ">=1.6.3,<2.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "3.3.3"
black = "23.11.0"
isort = "5.12.0"
flake8 = "6.1.0"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
